Massive inlining
This results in a further small speedup
in grid updates
